move segment name parsing to ref_builder.schema
