+++ ref_builder/otu/utils.py
@@ -5,29 +5,24 @@ from uuid import UUID, uuid4
 
 import structlog
 
-from ref_builder.models import Molecule, MolType
+from ref_builder.models import Molecule
 from ref_builder.ncbi.models import NCBIGenbank
-from ref_builder.schema import OTUSchema, SegmentName, Segment
+from ref_builder.schema import OTUSchema, Segment, get_multipartite_segment_name
 from ref_builder.utils import Accession, IsolateName, IsolateNameType
-
-
-SIMPLE_NAME_PATTERN = re.compile(r"([A-Za-z0-9])+")
-"""Regex pattern for parsing segment name strings with no prefix."""
-
-COMPLEX_NAME_PATTERN = re.compile(r"([A-Za-z]+)[-_ ]+([A-Za-z0-9]+)")
-"""Regex pattern for parsing segment name strings consisting of a prefix and a key."""
 
 logger = structlog.get_logger("otu.utils")
 
 
 class DeleteRationale(StrEnum):
     """Default strings delineating reasons for resource deletion."""
+
     USER = "Requested by user"
     REFSEQ = "Superceded by RefSeq"
 
 
 class RefSeqConflictError(ValueError):
     """Raised when a potential RefSeq replacement is found."""
+
     def __init__(
         self, message, isolate_id: UUID, isolate_name: IsolateName, accessions: list[str]
     ):
@@ -44,6 +39,7 @@ def create_schema_from_records(
     records: list[NCBIGenbank],
     segments: list[Segment] | None = None,
 ) -> OTUSchema | None:
+    """Return a complete schema from a list of records."""
     molecule = _get_molecule_from_records(records)
 
     binned_records = group_genbank_records_by_isolate(records)
@@ -65,7 +61,7 @@ def create_schema_from_records(
 def group_genbank_records_by_isolate(
     records: list[NCBIGenbank],
 ) -> dict[IsolateName, dict[Accession, NCBIGenbank]]:
-    """Indexes Genbank records by isolate name"""
+    """Indexes Genbank records by isolate name."""
     isolates = defaultdict(dict)
 
     for record in records:
@@ -74,8 +70,8 @@ def group_genbank_records_by_isolate(
 
             if isolate_name is None:
                 logger.debug(
-                    "RefSeq record does not contain sufficient source data for automatic inclusion."
-                    + "Add this record manually.",
+                    "RefSeq record does not contain sufficient source data "
+                    " for automatic inclusion. Add this record manually.",
                     accession=record.accession,
                     definition=record.definition,
                     source_data=record.source,
@@ -104,7 +100,7 @@ def parse_refseq_comment(comment: str) -> tuple[str, str]:
 
 
 def _get_molecule_from_records(records: list[NCBIGenbank]) -> Molecule:
-    """Return relevant molecule metadata from one or more records"""
+    """Return relevant molecule metadata from one or more records."""
     if not records:
         raise IndexError("No records given")
 
@@ -126,7 +122,7 @@ def _get_molecule_from_records(records: list[NCBIGenbank]) -> Molecule:
 
 
 def _get_isolate_name(record: NCBIGenbank) -> IsolateName | None:
-    """Get the isolate name from a Genbank record"""
+    """Get the isolate name from a Genbank record."""
     if record.source.model_fields_set.intersection(
         {IsolateNameType.ISOLATE, IsolateNameType.STRAIN, IsolateNameType.CLONE},
     ):
@@ -175,25 +171,3 @@ def _get_segments_from_records(records: list[NCBIGenbank]) -> list[Segment]:
             raise ValueError("No segment name found for multipartite OTU segment.")
 
     return segments
-
-
-def get_multipartite_segment_name(record: NCBIGenbank) -> SegmentName:
-    """Get a multipartite segment name from the record."""
-    if SIMPLE_NAME_PATTERN.fullmatch(record.source.segment):
-        return SegmentName(
-            prefix=record.moltype, key=record.source.segment,
-        )
-
-    return parse_segment_name(record.source.segment)
-
-
-def parse_segment_name(raw: str) -> SegmentName:
-    """Parse a SegmentName from a raw string."""
-    segment_name_parse = COMPLEX_NAME_PATTERN.fullmatch(raw)
-    if segment_name_parse:
-        return SegmentName(
-            prefix=segment_name_parse.group(1),
-            key=segment_name_parse.group(2)
-        )
-
-    raise ValueError(f"{raw} is not a valid segment name")
