+++ ref_builder/schema.py
@@ -1,8 +1,16 @@
+import re
+
 from pydantic import UUID4, BaseModel, computed_field
 from pydantic.dataclasses import dataclass
 
 from ref_builder.models import Molecule
-from ref_builder.ncbi.models import NCBISourceMolType
+from ref_builder.ncbi.models import NCBIGenbank, NCBISourceMolType
+
+SIMPLE_NAME_PATTERN = re.compile(r"([A-Za-z0-9])+")
+"""Regex pattern for parsing segment name strings with no prefix."""
+
+COMPLEX_NAME_PATTERN = re.compile(r"([A-Za-z]+)[-_ ]+([A-Za-z0-9]+)")
+"""Regex pattern for parsing segment name strings consisting of a prefix and a key."""
 
 
 @dataclass(frozen=True)
@@ -81,3 +89,25 @@ def determine_segment_prefix(moltype: NCBISourceMolType) -> str:
         return prefix
 
     raise ValueError(f"{moltype} may not be a valid NCBISourceMolType.")
+
+
+def parse_segment_name(raw: str) -> SegmentName:
+    """Parse a SegmentName from a raw string."""
+    segment_name_parse = COMPLEX_NAME_PATTERN.fullmatch(raw)
+    if segment_name_parse:
+        return SegmentName(
+            prefix=segment_name_parse.group(1),
+            key=segment_name_parse.group(2),
+        )
+
+    raise ValueError(f"{raw} is not a valid segment name")
+
+
+def get_multipartite_segment_name(record: NCBIGenbank) -> SegmentName:
+    """Get a multipartite segment name from the record."""
+    if SIMPLE_NAME_PATTERN.fullmatch(record.source.segment):
+        return SegmentName(
+            prefix=record.moltype, key=record.source.segment,
+        )
+
+    return parse_segment_name(record.source.segment)
